requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
openai>=1.12.0
httpx==0.25.2
numpy==1.26.2
python-dotenv==1.0.0
vaderSentiment
python-telegram-bot==20.7
apscheduler==3.10.4
python-dateutil==2.8.2
